Robustify the test
Synchronize the reads with the writes, so that the reader doesn't
de-chunk the entire body for us.  But this means that if flushing isn't
working properly, merely getting the response will cause us to lock up,
waiting for the writes to finish, which are waiting on the sync.  So
wrap the read in a goroutine, and fail if we don't finish it in a
reasonable amount of time.